move content of MatrixTools in Mat
