+++ Sofa/Component/SolidMechanics/FEM/Elastic/src/sofa/component/solidmechanics/fem/elastic/impl/MatrixTools.h
@@ -26,55 +26,4 @@
 namespace sofa::component::solidmechanics::fem::elastic
 {
 
-/**
- * Alias for a sofa::type::Mat
- * If T is sofa::type::Mat<L,C,real> and L==1 && C==1, the alias is the scalar type.
- * Otherwise, the alias is T itself.
- *
- * Example: static_cast<ScalarOrMatrix<MatType>>(matrix)
- */
-template <class T>
-using ScalarOrMatrix = std::conditional_t<
-    (T::nbLines==1 && T::nbCols==1), typename T::Real, T>;
-
-template <sofa::Size L, sofa::Size C, class real>
-real absGeneralizedDeterminant(const sofa::type::Mat<L, C, real>& mat)
-{
-    if constexpr (L == C)
-    {
-        return std::abs(sofa::type::determinant(mat));
-    }
-    else
-    {
-        return std::sqrt(sofa::type::determinant(mat.multTranspose(mat)));
-    }
-}
-
-template <sofa::Size L, sofa::Size C, class real>
-sofa::type::Mat<C, L, real> leftPseudoInverse(const sofa::type::Mat<L, C, real>& mat)
-{
-    return mat.multTranspose(mat).inverted() * mat.transposed();
-}
-
-/**
- * Computes the inverse of a given matrix.
- * For square matrices (L == C), the standard matrix inverse is computed.
- * For non-square matrices, the left pseudo-inverse is returned.
- *
- * @param mat The input matrix of size LxC to be inverted or pseudo-inverted.
- * @return A matrix of size CxL representing the inverse or left pseudo-inverse of the input matrix.
- */
-template <sofa::Size L, sofa::Size C, class real>
-sofa::type::Mat<C, L, real> inverse(const sofa::type::Mat<L, C, real>& mat)
-{
-    if constexpr (L == C)
-    {
-        return mat.inverted();
-    }
-    else
-    {
-        return leftPseudoInverse(mat);
-    }
-}
-
 }
